Update scraper to be faster and more classy
Instead of going from "A* A*" -> "AA* AA*" to handle the "too many
results" error this goes from "A* A*" to just "A* AA*", which makes
things about 26x faster.

Also this puts the whole scraper in a class for purposes of retrying.
Also this removes the intermediate "[x] results" messages since they
complicate the recursion, and it's better to just output the results as
they come in.